Add logback.xml to configure logging related to GWT compilation
First of all, this limits Errai logging level from DEBUG to INFO which
significantly reduces output size. This is important when analyzing
console output on Jenkins.

It should be possible to adjust Errai logging level using a system
property for debugging purpose but I wasn't able to get it working
locally. The syntax looks as expected and it used to work like this
in OptaPlanner Examples. We'll figure this out later.