RHDM-257: Avoid CDI warning
WELD-000167: Class SolverValidator$1 is annotated with @ApplicationScoped
but it does not declare an appropriate constructor therefore is not
registered as a bean!

The inner class slightly alters behavior of DefaultGenericKieValidator by
overriding one of its methods. As a side effect it becomes a bean
because @ApplicationScoped is inherited.

The solution is to exclude the inner class from bean scanning as it's
never being used as a bean.